Restructure the decode module
